backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
groq
python-dotenv
psycopg2-binary
